happy path for DePool: extract a helper for deploying the setup
